migration from memory to jdbc and mysql

+++ src/main/java/service/FinanceService.java
@@ -3,6 +3,6 @@ package service;
 import model.Account;
 
 public interface FinanceService {
-    void deposit(Account account, long amount);
-    void withdraw(Account account, long amount);
+    void deposit(Account account, int amount);
+    void withdraw(Account account, int amount);
 }
